Fix float64 error
When read motor's angle, use np array with float32. Default makes float64

+++ lerobot/common/robot_devices/motors/lewansoul.py
@@ -330,7 +330,7 @@ class LewansoulMotorsBus:
             values.append(value)
 
         print(values)
-        values = np.array(values)
+        values = np.array(values, dtype=np.float32)
         return values
     
     def write(self, data_name, values: int | float | np.ndarray, motor_names: str | list[str] | None = None):
